requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.1.3
django-allauth==65.9.0
django-axes==8.0.0
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
django-js-asset==3.1.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
kombu==5.5.4
MetaTrader5==5.0.4993
numpy==2.2.5
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.2
Telethon==1.40.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
